# backend/openapi.yaml
openapi: 3.0.3
info:
  title: riverside-lite API
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Participant:
      type: object
      required: [id, recordingId, role]
      properties:
        id: { type: string, format: uuid }
        recordingId: { type: string, format: uuid }
        role: { type: string, enum: [host, guest] }
        displayName: { type: string }
        email: { type: string, format: email }
...
    AddParticipantRequest:
      type: object
      required: [displayName, role]
      properties:
        displayName: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [host, guest] }
    AddParticipantResponse:
      type: object
      properties:
        participant: { $ref: '#/components/schemas/Participant' }
        magicLink: { type: string, description: 'One-time link for guest' }
    UploadInitiateRequest:
      type: object
      required: [recordingId, participantId, kind, protocol]
      properties:
        recordingId: { type: string, format: uuid }
        participantId: { type: string, format: uuid }
        kind: { type: string, enum: [audio, video] }
        protocol: { type: string, enum: [tus, multipart] }
        mimeType: { type: string }
        codec: { type: string }
    UploadInitiateResponse:
      type: object
      required: [uploadId, protocol]
      properties:
        uploadId: { type: string }
        protocol: { type: string, enum: [tus, multipart] }
        tus:
          type: object
          properties:
            uploadUrl: { type: string, description: 'tus creation URL (POST) or resource URL' }
        multipart:
          type: object
          properties:
            sasUrl: { type: string, description: 'SAS base URL for Put Block (Azure) or presigned URL (AWS S3 part)' }
            partSize: { type: integer }
    UploadCompleteRequest:
      type: object
      required: [uploadId]
      properties:
        uploadId: { type: string }
        sizeBytes: { type: integer }
        etag: { type: string }
...
paths:
...
  /v1/recordings/{id}/participants:
    post:
      summary: Add participant (returns magic link for guests)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddParticipantRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AddParticipantResponse' }
  /v1/uploads/initiate:
    post:
      summary: Start resumable upload (tus or multipart)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UploadInitiateRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadInitiateResponse' }
  /v1/uploads/{id}/complete:
    post:
      summary: Commit uploaded track and trigger processing
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UploadCompleteRequest' }
      responses:
        '202': { description: Accepted }
...